feat: delete user himslef, last responsive fix

+++ backend/controllers/post.controller.js
@@ -6,6 +6,7 @@ const mongoose = require('mongoose')
 
 module.exports.getPost = (req,res,next)=>{
     postModel.find((err,docs)=>{
+        // postModel.imageUrl = `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
         if (!err) res.send(docs)
         else{
             console.log('error to get data : ' + err)
@@ -25,24 +26,20 @@ exports.addPost = async (req,res,next)=>{
         likes:0,
         userLiked:[],
     })
-    console.log('req file'+req.file)
     if (req.file != null || req.file != undefined) {
         post.imageUrl = `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
         console.log(post.imageUrl)
     }
-    console.log("post", post)
     post.save()
         .then(()=>{
             res.status(200).json({message:"enregistrement du post réussi !"})
           })
           .catch((error)=>{
             res.status(420).json({error})
-            console.log('dans le catch')
           })
 }
 
 module.exports.modifyPost = (req,res,next)=>{
-    console.log("req file", req.file)
 
     const newPost = JSON.parse(req.body.post);
 
@@ -88,17 +85,22 @@ module.exports.modifyPost = (req,res,next)=>{
 exports.deletePost = (req, res, next) => {
     // Find the sauce by id
     postModel.findOne({ _id: req.params.id})
-        .then(post => {
-            // retrieve filename by path of img
-            
-                const filename = post.imageUrl.split('/images/')[1]
-
+        .then(post => {        
                 // Delete the img of post removed
-                fs.unlink(`images/${filename}`, () => {
+                if(post.imageUrl){
+                    // retrieve filename by path of img    
+                    const filename = post.imageUrl.split('/images/')[1] 
+                    fs.unlink(`images/${filename}`, () => {
+                        postModel.deleteOne({_id: req.params.id})
+                            .then(() => { res.status(200).json({message: 'post supprimé !'})})
+                            .catch(error => res.status(401).json({ error }))
+                    });
+                }else{
                     postModel.deleteOne({_id: req.params.id})
-                        .then(() => { res.status(200).json({message: 'post supprimé !'})})
-                        .catch(error => res.status(401).json({ error }))
-                });
+                            .then(() => { res.status(200).json({message: 'post supprimé !'})})
+                            .catch(error => res.status(401).json({ error }))
+                }
+                
             })
         .catch( error => {
             res.status(500).json({ error });
